% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCondition.R
\name{buildCondition}
\alias{buildCondition}
\title{Builds a condition}
\usage{
buildCondition(variable, operator, value)
}
\arguments{
\item{variable}{The variable or column name to filter.}

\item{operator}{An operator for the filter. Currenty (== equals) an (!= not equals) are supported. (More operators to be implemented in the future).}

\item{value}{Provides a value to compare with.}
}
\value{
Returns the filtering clause. A list of clauses can be composed
and passed to \code{query()} or \code{count()} functions via the \code{conditions}
parameter to build complex queries.
}
\description{
Returns a condition clause to be used in queries.
}
\examples{
cnx <- connect("http://jacaton-r.herokuapp.com", "demo", "1234") 
off <- resource(cnx, "oficinas") 

count(off)
count(off, conditions=buildCondition("nombre", "!=", "Seville")  )
count(off, conditions=buildCondition("nombre", "==", "Seville")  )
query(off, conditions=buildCondition("nombre", "==", "Seville")  )
}

