% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(url, user, pass)
}
\arguments{
\item{url}{The base URL to connect to.}

\item{user}{The user credential.}

\item{pass}{The password for the user.}
}
\value{
connect returns a connection object. This connection object can be used to discover and access to resources.
}
\description{
Setup a connection to a given Hivepod backend with credentials.
}
\examples{
connect("http://jacaton-r.herokuapp.com", "demo", "1234") 
}

