% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Creates data in the Hivepod backend.}
\usage{
create(resource, rows)
}
\arguments{
\item{resource}{The resource to work with. Returned by the \code{resource()} function.}

\item{rows}{Rows to create.}
}
\value{
Saved data. 
For new created data Hivepod will setup a  value for the \code{_id} field.
}
\description{
Creates  data in the cloud backend.
}

