% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Deletes data in the Hivepod backend.}
\usage{
delete(resource, conditions = NULL, id = NULL, ids = NULL, rows = NULL)
}
\arguments{
\item{resource}{The resource to work with. Returned by the \code{resource()} function.}

\item{conditions}{Conditions to delete all data matching the query.}

\item{id}{If of the data row to delete.}

\item{ids}{List of ids to remove.}

\item{rows}{Rows to remove.}
}
\value{
The deleted data. Use only one criteria for deletion: conditions or id or ids or rows.
}
\description{
Deletes data in the cloud backend by id, ids or rows.
}
\examples{

}

