% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{lo que hace la funcion}
\usage{
hello()
}
\arguments{
\item{parametro}{1 \code{}}

\item{param}{2}
}
\value{
Ea0 returns a data.frame of class renyi with the selected indices.

EinvD returns a vector with EinvD.

EJ returns a a vector with Pielou's index.

Evar returns a vector with Evar.
}
\description{
en detalle
}
\examples{
Ea0(A = dummy$abun)
Ea0(A = dummy$abun, scales = c(0.25,0.5,1,2,4,8,Inf))
EinvD(A = dummy$abun)
EJ(A = dummy$abun)
Evar(A = dummy$abun)
#calculate all of them
eve1 <- Eve(A = dummy$abun)
eve2 <- Eve(A = dummy$abun, scales = c(0.25,0.5,1,2,4,8,Inf))
pairs(eve1)
pairs(eve2)
}

