% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{Saves data in the Hivepod backend.}
\usage{
save(resource, rows)
}
\arguments{
\item{resource}{The resource to work with. Returned by the \code{resource()} function.}

\item{rows}{Rows to save.}
}
\value{
Saved data. 
New data with no defined \code{_id} property will be created as new. POST operation will be atempted.
Preexisting records with a value in \code{_id} property will be updated in the backend. PUT operation will be attempted.
}
\description{
Saves data in the cloud backend.
}
\examples{
cnx <- connect("http://jacaton-r.herokuapp.com", "demo", "1234") 
not <- resource(cnx, "noticias") 
res <- save(not, list(titular="a", fecha="2016-02-05T12:56:00.345Z", cuerpo="body")) 
}

