% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dataframe.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Converts returned JSON data into an R dataframe.}
\usage{
to_dataframe(response, keep_metadata = "clean")
}
\arguments{
\item{response}{The raw response data after a query.}

\item{keep_metadata}{Flag to indicate if metadatata should be handled.
Suported values are: 
   - "clean"     -  remove metadata (default one)
   - "raw"       -  remove metadata
   - "raw_links" -  remove metadata links}
}
\value{
\code{resource()} returns a resource object for a Hivepod backend. 
This resource object can be used to make queries, count and manipulate data in the backend.
}
\description{
Converts returned JSON data into an R dataframe.
}
\examples{
cnx <- connect("http://jacaton-r.herokuapp.com", "demo", "1234") 
exo <- resource(cnx, "exoplanets") 
whe <- resource(cnx, "wheathers") 
oly <- resource(cnx, "olympicMedals") 
}

