% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Updates data in the Hivepod backend.}
\usage{
update(resource, rows)
}
\arguments{
\item{resource}{The resource to work with. Returned by the \code{resource()} function.}

\item{rows}{Rows to update (must contains the \code{_id} field).}
}
\value{
Updated data.
}
\description{
Updates data in the cloud backend.
}

