% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{urlEncode}
\alias{urlEncode}
\title{Encode a URLComponent}
\usage{
urlEncode(value)
}
\arguments{
\item{value}{The value to encode.}
}
\value{
Returns a value protected for URL encoding. 
Add quotes if value is a string. 
Convert R logical value to JSON booleans.
}
\description{
Private function. Returns a value protected for URL encoding.
}
\examples{
urlEncode("A B") 
output> "A\%20B"
}

